requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
fastapi==0.110.1
fastapi-cache2==0.2.1
flower==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
kombu==5.3.6
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
pendulum==3.0.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-stdnum==1.20
pytz==2024.1
redis==5.0.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
time-machine==2.14.1
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
